Move schemas to its own place
