+++ client/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod schemas;
 pub mod screens;
 pub mod theme;
