+++ client/src/schemas.rs
@@ -1,0 +1,3 @@
+// Objects to deal with the Rest API
+
+pub mod deck;
