+++ client/src/screens/team.rs
@@ -1,73 +1,19 @@
-use std::fmt::{self, Display};
-
 use cursive::views::{Dialog, ListView, ScrollView, TextView};
 use cursive::Cursive;
 use reqwest::blocking::Client;
-use serde::{Deserialize, Serialize};
 
-use super::MainMenu;
-
-#[derive(Debug, Clone)]
-#[derive(Serialize, Deserialize)]
-#[derive(PartialEq, Eq, Hash)]
-pub struct Player {
-    name: String,
-    position: String,
-    attack: u32,
-    defense: u32,
-    passing: u32,
-    stamina: u32,
-}
-
-impl Display for Player {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{} ({}) | ATK: {} | DEF: {} | PASS: {} | STA: {}",
-            self.name, self.position, self.attack, self.defense, self.passing, self.stamina
-        )
-    }
-}
-
-#[derive(Debug, Clone)]
-#[derive(Serialize, Deserialize)]
-#[derive(PartialEq, Eq, Hash)]
-pub struct PowerUp {
-    name: String,
-    effect: String,
-}
-
-impl Display for PowerUp {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}: {}", self.name, self.effect)
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub struct DeckResponse {
-    pub players: Vec<Player>,
-    pub power_ups: Vec<(PowerUp, u32)>,
-
-    #[allow(dead_code, reason = "needed for deserialization purposes.")]
-    message: String,
-}
+use crate::screens;
+use crate::schemas::deck::DeckResponse;
+use crate::schemas::deck::Player;
+use crate::schemas::deck::PowerUp;
 
 #[allow(non_snake_case)]
 pub fn TeamScreen(app: &mut Cursive, auth: String) {
     app.pop_layer();
 
-    let deck: DeckResponse = match user_deck(&auth) {
-        Ok(resp) => match resp.json() {
-            Ok(json) => json,
-            Err(_) => {
-                app.add_layer(Dialog::info("Failed to parse deck JSON"));
-                return;
-            }
-        },
-        Err(_) => {
-            app.add_layer(Dialog::info("Failed to fetch deck"));
-            return;
-        }
+    let deck = match deck_of(app, &auth) {
+        Some(value) => value,
+        None => return,
     };
 
     let mut list = ListView::new();
@@ -84,16 +30,30 @@ pub fn TeamScreen(app: &mut Cursive, auth: String) {
     app.add_layer(
         Dialog::around(ScrollView::new(list).scroll_y(true))
             .title("Team Screen")
-            .button("Back to Main", move |s| MainMenu(s, auth_clone.clone())),
+            .button("Back to Main", move |s| screens::MainMenu(s, auth_clone.clone())),
     );
+}
+
+fn deck_of(app: &mut Cursive, auth: &String) -> Option<DeckResponse> {
+    let deck: DeckResponse = match user_deck(auth) {
+        Ok(resp) => match resp.json() {
+            Ok(json) => json,
+            Err(_) => {
+                app.add_layer(Dialog::info("Failed to parse deck JSON"));
+                return None;
+            }
+        },
+        Err(_) => {
+            app.add_layer(Dialog::info("Failed to fetch deck"));
+            return None;
+        }
+    };
+    Some(deck)
 }
 
 #[allow(non_snake_case)]
 fn CardItem(i: usize, player: &Player, auth_clone: String) -> Dialog {
-    let player_info = format!(
-        "{} - {} | ATK: {} DEF: {} PASS: {} STA: {}",
-        player.position, player.name, player.attack, player.defense, player.passing, player.stamina
-    );
+    let player_info = format!("{}", player);
 
     Dialog::around(TextView::new(player_info)).button("Remove", move |s| {
         fire_player(i, &auth_clone);
@@ -104,7 +64,7 @@ fn CardItem(i: usize, player: &Player, auth_clone: String) -> Dialog {
 
 #[allow(non_snake_case)]
 fn PowerItem(power: &PowerUp, count: &u32) -> TextView {
-    let power_info = format!("{} x{} - {}", power.name, count, power.effect);
+    let power_info = format!("{}x : {}", count, power);
     TextView::new(power_info)
 }
